+++ changelog.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.0.2] - 2021-05-06
+
+### Added
+
+* `start` command: open cloned project in a specified editor
+  * `-e/--editor` and `-n/--no-open` arguments
+
+### Fixed
+
+* Small script help fixes/improvements
+
+
 ## [0.0.1] - 2021-05-05
 
 ### Added